⬆️ migrate app module to standalone
