+++ src/environments/environment.production.ts
@@ -2,4 +2,5 @@ import { EnvironmentType } from './environment.type';
 
 export const environment: EnvironmentType = {
   production: true,
+  environmentSpecificModules: [],
 };
